Working live views in unprivileged sections
