+++ apps/fz_http/assets/js/unprivileged.js
@@ -7,7 +7,7 @@ import "@fontsource/open-sans"
 import "@fontsource/fira-mono"
 
 import "phoenix_html"
+import "./live_view.js"
 import "./event_listeners.js"
-import { fzCrypto } from "./crypto.js"
 
-window.fzCrypto = fzCrypto
+import { fzCrypto } from "./crypto.js"
