+++ apps/fz_http/lib/fz_http_web/live/device_live/admin/show.html.heex
@@ -1,6 +1,6 @@
 <%= if @live_action == :edit do %>
   <%= live_modal(
-        FzHttpWeb.DeviceLive.UpdateFormComponent,
+        FzHttpWeb.DeviceLive.Admin.UpdateFormComponent,
         return_to: Routes.device_admin_show_path(@socket, :show, @device),
         title: "Edit #{@device.name}",
         id: "device-#{@device.id}",
