+++ apps/fz_http/lib/fz_http_web/live/device_live/admin/update_form_component.ex
@@ -1,4 +1,4 @@
-defmodule FzHttpWeb.DeviceLive.UpdateFormComponent do
+defmodule FzHttpWeb.DeviceLive.Admin.UpdateFormComponent do
   @moduledoc """
   Handles device form.
   """
