+++ apps/fz_http/lib/fz_http_web/live/device_live/unprivileged/index.html.heex
@@ -1,23 +1,68 @@
 <%= if @live_action == :new do %>
   <%= live_modal(
-        FzHttpWeb.DeviceLive.CreateFormComponent,
-        return_to: Routes.device_unprivileged_index_path(@socket, :index),
-        title: "Add Device",
-        id: "add-device-modal",
-        target_user_id: @current_user.id,
-        action: @live_action) %>
+      FzHttpWeb.DeviceLive.Unprivileged.NewComponent,
+      return_to: Routes.device_unprivileged_index_path(@socket, :index),
+      title: "Add WireGuard Tunnel",
+      action: @live_action,
+      id: "new-tunnel-component") %>
 <% end %>
 
-<%= render FzHttpWeb.SharedView, "heading.html", page_title: @page_title %>
-
 <section class="section is-main-section">
+  <h4 class="title is-4"><%= @page_title %></h4>
+
+  <div class="block">
+    Each tunnel represents a WireGuard configuration for connecting to this
+    Firezone server.
+  </div>
+
   <%= render FzHttpWeb.SharedView, "flash.html", assigns %>
 
   <div class="block">
-    <%= render FzHttpWeb.SharedView, "devices_table.html", devices: @devices, socket: @socket %>
+    <%= if length(@devices) > 0 do %>
+      <table class="table is-bordered is-hoverable is-striped is-fullwidth">
+        <thead>
+          <tr>
+            <th>Name</th>
+            <th>Assigned Tunnel IP</th>
+            <th>Public key</th>
+            <th>Created</th>
+            <th></th>
+          </tr>
+        </thead>
+        <tbody>
+        <%= for device <- @devices do %>
+          <tr>
+            <td><%= device.name %></td>
+            <td class="code">
+              <span><%= device.ipv4 %></span>
+              <span><%= device.ipv6 %></span>
+            </td>
+            <td class="code"><%= device.public_key %></td>
+            <td id={"device-#{device.id}-inserted-at"} data-timestamp={device.inserted_at} phx-hook="FormatTimestamp">…</td>
+            <td>
+              <a href="#" phx-click="delete" phx-val-id={device.id}>Delete</a>
+            </td>
+          </tr>
+        <% end %>
+        </tbody>
+      </table>
+    <% else %>
+      <p>
+        <strong>No WireGuard tunnels to display.</strong>
+      </p>
+    <% end %>
   </div>
 
-  <%= live_patch(to: Routes.device_unprivileged_index_path(@socket, :new), class: "button") do %>
-    Generate New Configuration
-  <% end %>
+  <div class="level">
+    <div class="level-left">
+      <%= live_patch(to: Routes.device_unprivileged_index_path(@socket, :new), class: "button") do %>
+        Add a Tunnel
+      <% end %>
+    </div>
+    <div class="level-right">
+      <%= link(to: Routes.session_path(@socket, :delete), method: :delete) do %>
+        Sign out
+      <% end %>
+    </div>
+  </div>
 </section>
