+++ apps/fz_http/lib/fz_http_web/templates/layout/admin.html.heex
@@ -16,15 +16,7 @@
     <meta name="msapplication-config" content="/browserconfig.xml" />
     <meta name="msapplication-TileColor" content="331700">
     <meta name="theme-color" content="331700">
-
-    <!-- User Socket -->
-    <%= tag :meta, name: "user-token", content: Phoenix.Token.sign(@conn, "user auth", @current_user.id) %>
-
-    <!-- Notification Channel -->
-    <%= tag :meta, name: "channel-token", content: Phoenix.Token.sign(@conn, "channel auth", @current_user.id) %>
-
-    <!-- CSRF -->
-    <%= csrf_meta_tag() %>
+    <%= render(FzHttpWeb.SharedView, "socket_token_headers.html", conn: @conn, current_user: @current_user) %>
   </head>
   <body>
 <div id="app">
