--- apps/fz_http/lib/fz_http_web/templates/layout/email.html.eex
@@ -1,8 +1,0 @@
-<html>
-  <head>
-    <link rel="stylesheet" href="<%= Routes.static_path(FzHttpWeb.Endpoint, "/css/email.css") %>">
-  </head>
-  <body>
-    <%= @inner_content %>
-  </body>
-</html>
