+++ apps/fz_http/lib/fz_http_web/templates/layout/email.html.heex
@@ -1,0 +1,8 @@
+<html>
+  <head>
+    <link rel="stylesheet" href={Routes.static_path(FzHttpWeb.Endpoint, "/css/email.css")}>
+  </head>
+  <body>
+    <%= @inner_content %>
+  </body>
+</html>
