+++ apps/fz_http/lib/fz_http_web/templates/layout/unprivileged.html.heex
@@ -4,7 +4,6 @@
     <meta charset="utf-8"/>
     <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <%= csrf_meta_tag() %>
     <%= live_title_tag assigns[:page_title] || "Firezone" %>
     <link phx-track-static rel="stylesheet" href={Routes.static_path(@conn, "/css/app.css")} />
     <script defer phx-track-static type="text/javascript" src={Routes.static_path(@conn, "/js/unprivileged.js")}></script>
@@ -17,19 +16,22 @@
     <meta name="msapplication-config" content="/browserconfig.xml" />
     <meta name="msapplication-TileColor" content="331700">
     <meta name="theme-color" content="331700">
+    <%= render(FzHttpWeb.SharedView, "socket_token_headers.html", current_user: @current_user, conn: @conn) %>
   </head>
 <body>
 <section class="section hero is-fullheight is-error-section">
   <div id="app" class="hero-body">
     <div class="container">
       <div class="columns is-centered">
-        <div class="column is-two-thirds-tablet is-half-desktop is-one-third-widescreen">
+        <div class="column">
           <div class="block">
             <div class="has-text-centered">
               <img src={Routes.static_path(@conn, "/images/logo-text.svg")} alt="firez.one">
             </div>
           </div>
+
           <%= @inner_content %>
+
         </div>
       </div>
     </div>
